[lldb-dap] Refactor breakpoint related request handlers (NFC)
